Test cleanup: Move config-related utils to config package
This commit moves the remaining config-related utils in tests/utils.go to a specific config package.
