+++ tests/compute/BUILD.bazel
@@ -24,6 +24,7 @@ go_library(
         "//tests/framework/kubevirt:go_default_library",
         "//tests/framework/matcher:go_default_library",
         "//tests/libkubevirt:go_default_library",
+        "//tests/libkubevirt/config:go_default_library",
         "//tests/libnet/cloudinit:go_default_library",
         "//tests/libsecret:go_default_library",
         "//tests/libvmifact:go_default_library",
