+++ tests/compute/vmidefaults.go
@@ -38,6 +38,7 @@ import (
 	"kubevirt.io/kubevirt/tests/framework/kubevirt"
 	"kubevirt.io/kubevirt/tests/framework/matcher"
 	"kubevirt.io/kubevirt/tests/libkubevirt"
+	"kubevirt.io/kubevirt/tests/libkubevirt/config"
 	"kubevirt.io/kubevirt/tests/libvmifact"
 	"kubevirt.io/kubevirt/tests/libwait"
 	"kubevirt.io/kubevirt/tests/testsuite"
@@ -149,7 +150,7 @@ var _ = SIGDescribe("VMIDefaults", func() {
 			By("Adding period to virt-config")
 			kvConfigurationCopy := kvConfiguration.DeepCopy()
 			kvConfigurationCopy.MemBalloonStatsPeriod = &period
-			tests.UpdateKubeVirtConfigValueAndWait(*kvConfigurationCopy)
+			config.UpdateKubeVirtConfigValueAndWait(*kvConfigurationCopy)
 			if kvConfiguration.VirtualMachineOptions != nil && kvConfiguration.VirtualMachineOptions.DisableFreePageReporting != nil {
 				expected.FreePageReporting = "off"
 			} else {
