+++ tests/instancetype/BUILD.bazel
@@ -22,7 +22,6 @@ go_library(
         "//staging/src/kubevirt.io/api/instancetype/v1beta1:go_default_library",
         "//staging/src/kubevirt.io/client-go/kubecli:go_default_library",
         "//staging/src/kubevirt.io/client-go/kubevirt/scheme:go_default_library",
-        "//tests:go_default_library",
         "//tests/decorators:go_default_library",
         "//tests/framework/checks:go_default_library",
         "//tests/framework/kubevirt:go_default_library",
@@ -33,6 +32,7 @@ go_library(
         "//tests/libinstancetype/builder/v1alpha1:go_default_library",
         "//tests/libinstancetype/builder/v1alpha2:go_default_library",
         "//tests/libkubevirt:go_default_library",
+        "//tests/libkubevirt/config:go_default_library",
         "//tests/libstorage:go_default_library",
         "//tests/libvmifact:go_default_library",
         "//tests/libvmops:go_default_library",
