+++ tests/monitoring/BUILD.bazel
@@ -20,7 +20,6 @@ go_library(
         "//pkg/monitoring/metrics/virt-operator:go_default_library",
         "//staging/src/kubevirt.io/api/core/v1:go_default_library",
         "//staging/src/kubevirt.io/client-go/kubecli:go_default_library",
-        "//tests:go_default_library",
         "//tests/clientcmd:go_default_library",
         "//tests/console:go_default_library",
         "//tests/decorators:go_default_library",
@@ -30,6 +29,7 @@ go_library(
         "//tests/framework/kubevirt:go_default_library",
         "//tests/framework/matcher:go_default_library",
         "//tests/libkubevirt:go_default_library",
+        "//tests/libkubevirt/config:go_default_library",
         "//tests/libmigration:go_default_library",
         "//tests/libmonitoring:go_default_library",
         "//tests/libnet:go_default_library",
