+++ tests/virtiofs/datavolume.go
@@ -41,7 +41,6 @@ import (
 	"kubevirt.io/kubevirt/pkg/libvmi"
 	libvmici "kubevirt.io/kubevirt/pkg/libvmi/cloudinit"
 	virtconfig "kubevirt.io/kubevirt/pkg/virt-config"
-	"kubevirt.io/kubevirt/tests"
 	"kubevirt.io/kubevirt/tests/console"
 	cd "kubevirt.io/kubevirt/tests/containerdisk"
 	"kubevirt.io/kubevirt/tests/decorators"
@@ -51,7 +50,7 @@ import (
 	. "kubevirt.io/kubevirt/tests/framework/matcher"
 	"kubevirt.io/kubevirt/tests/libdv"
 	"kubevirt.io/kubevirt/tests/libkubevirt"
-	"kubevirt.io/kubevirt/tests/libkubevirt/config"
+	kvconfig "kubevirt.io/kubevirt/tests/libkubevirt/config"
 	"kubevirt.io/kubevirt/tests/libpod"
 	"kubevirt.io/kubevirt/tests/libstorage"
 	"kubevirt.io/kubevirt/tests/libvmifact"
@@ -87,9 +86,9 @@ var _ = Describe("[sig-storage] virtiofs", decorators.SigStorage, func() {
 
 		DescribeTable("[Serial] should be successfully started and accessible", Serial, func(namespace string) {
 			if namespace == testsuite.NamespacePrivileged {
-				config.EnableFeatureGate(virtconfig.VirtIOFSGate)
+				kvconfig.EnableFeatureGate(virtconfig.VirtIOFSGate)
 			} else {
-				config.DisableFeatureGate(virtconfig.VirtIOFSGate)
+				kvconfig.DisableFeatureGate(virtconfig.VirtIOFSGate)
 			}
 
 			createPVC(namespace, pvc1)
@@ -165,7 +164,7 @@ var _ = Describe("[sig-storage] virtiofs", decorators.SigStorage, func() {
 		})
 
 		AfterEach(func() {
-			tests.UpdateKubeVirtConfigValueAndWait(originalConfig)
+			kvconfig.UpdateKubeVirtConfigValueAndWait(originalConfig)
 		})
 
 		createHostPathPV := func(pvc, namespace string) {
@@ -208,11 +207,11 @@ var _ = Describe("[sig-storage] virtiofs", decorators.SigStorage, func() {
 					Resources: resources,
 				},
 			}
-			tests.UpdateKubeVirtConfigValueAndWait(*config)
+			kvconfig.UpdateKubeVirtConfigValueAndWait(*config)
 			if namespace == testsuite.NamespacePrivileged {
-				config.EnableFeatureGate(virtconfig.VirtIOFSGate)
+				kvconfig.EnableFeatureGate(virtconfig.VirtIOFSGate)
 			} else {
-				config.DisableFeatureGate(virtconfig.VirtIOFSGate)
+				kvconfig.DisableFeatureGate(virtconfig.VirtIOFSGate)
 			}
 
 			pvcName := fmt.Sprintf("disk-%s", pvc)
@@ -286,9 +285,9 @@ var _ = Describe("[sig-storage] virtiofs", decorators.SigStorage, func() {
 
 		DescribeTable("[Serial] should be successfully started and virtiofs could be accessed", Serial, func(namespace string) {
 			if namespace == testsuite.NamespacePrivileged {
-				config.EnableFeatureGate(virtconfig.VirtIOFSGate)
+				kvconfig.EnableFeatureGate(virtconfig.VirtIOFSGate)
 			} else {
-				config.DisableFeatureGate(virtconfig.VirtIOFSGate)
+				kvconfig.DisableFeatureGate(virtconfig.VirtIOFSGate)
 			}
 
 			dataVolume := libdv.NewDataVolume(
