+++ tests/vmi_configuration_test.go
@@ -69,7 +69,7 @@ import (
 	. "kubevirt.io/kubevirt/tests/framework/matcher"
 	"kubevirt.io/kubevirt/tests/libdv"
 	"kubevirt.io/kubevirt/tests/libkubevirt"
-	"kubevirt.io/kubevirt/tests/libkubevirt/config"
+	kvconfig "kubevirt.io/kubevirt/tests/libkubevirt/config"
 	"kubevirt.io/kubevirt/tests/libnet"
 	"kubevirt.io/kubevirt/tests/libnet/cloudinit"
 	"kubevirt.io/kubevirt/tests/libnode"
@@ -551,7 +551,7 @@ var _ = Describe("[sig-compute]Configurations", decorators.SigCompute, func() {
 
 				config := kv.Spec.Configuration
 				config.DeveloperConfiguration.MemoryOvercommit = 200
-				tests.UpdateKubeVirtConfigValueAndWait(config)
+				kvconfig.UpdateKubeVirtConfigValueAndWait(config)
 			})
 
 			It("[test_id:3114]should set requested amount of memory according to the specified virtual memory", func() {
@@ -1325,7 +1325,7 @@ var _ = Describe("[sig-compute]Configurations", decorators.SigCompute, func() {
 
 					config := kv.Spec.Configuration
 					config.SupportedGuestAgentVersions = []string{"X.*"}
-					tests.UpdateKubeVirtConfigValueAndWait(config)
+					kvconfig.UpdateKubeVirtConfigValueAndWait(config)
 				})
 
 				It("[test_id:5267]VMI condition should signal unsupported agent presence", func() {
@@ -1642,7 +1642,7 @@ var _ = Describe("[sig-compute]Configurations", decorators.SigCompute, func() {
 				By("Configuring a default runtime class")
 				config := libkubevirt.GetCurrentKv(virtClient).Spec.Configuration.DeepCopy()
 				config.DefaultRuntimeClass = runtimeClassName
-				tests.UpdateKubeVirtConfigValueAndWait(*config)
+				kvconfig.UpdateKubeVirtConfigValueAndWait(*config)
 
 				By("Creating a new VMI")
 				vmi := libvmifact.NewGuestless()
@@ -1676,7 +1676,7 @@ var _ = Describe("[sig-compute]Configurations", decorators.SigCompute, func() {
 
 				config := kv.Spec.Configuration
 				config.AdditionalGuestMemoryOverheadRatio = &ratioStr
-				tests.UpdateKubeVirtConfigValueAndWait(config)
+				kvconfig.UpdateKubeVirtConfigValueAndWait(config)
 			}
 
 			getComputeMemoryRequest := func(vmi *v1.VirtualMachineInstance) resource.Quantity {
@@ -1875,7 +1875,7 @@ var _ = Describe("[sig-compute]Configurations", decorators.SigCompute, func() {
 			config.ArchitectureConfiguration.Arm64.EmulatedMachines = testEmulatedMachines
 			config.ArchitectureConfiguration.Ppc64le.EmulatedMachines = testEmulatedMachines
 
-			tests.UpdateKubeVirtConfigValueAndWait(config)
+			kvconfig.UpdateKubeVirtConfigValueAndWait(config)
 		})
 
 		It("[test_id:3124]should set machine type from VMI spec", func() {
@@ -1930,7 +1930,7 @@ var _ = Describe("[sig-compute]Configurations", decorators.SigCompute, func() {
 			config.ArchitectureConfiguration.Amd64.EmulatedMachines = testEmulatedMachines
 			config.ArchitectureConfiguration.Arm64.EmulatedMachines = testEmulatedMachines
 			config.ArchitectureConfiguration.Ppc64le.EmulatedMachines = testEmulatedMachines
-			tests.UpdateKubeVirtConfigValueAndWait(config)
+			kvconfig.UpdateKubeVirtConfigValueAndWait(config)
 
 			vmi := libvmifact.NewGuestless()
 			vmi = libvmops.RunVMIAndExpectLaunch(vmi, 30)
@@ -1987,7 +1987,7 @@ var _ = Describe("[sig-compute]Configurations", decorators.SigCompute, func() {
 			config := kv.Spec.Configuration
 			configureCPURequest := resource.MustParse("800m")
 			config.CPURequest = &configureCPURequest
-			tests.UpdateKubeVirtConfigValueAndWait(config)
+			kvconfig.UpdateKubeVirtConfigValueAndWait(config)
 
 			vmi := libvmifact.NewGuestless()
 			runningVMI := libvmops.RunVMIAndExpectScheduling(vmi, 30)
@@ -2009,7 +2009,7 @@ var _ = Describe("[sig-compute]Configurations", decorators.SigCompute, func() {
 			config.AutoCPULimitNamespaceLabelSelector = &metav1.LabelSelector{
 				MatchLabels: map[string]string{autoCPULimitLabel: "true"},
 			}
-			tests.UpdateKubeVirtConfigValueAndWait(config)
+			kvconfig.UpdateKubeVirtConfigValueAndWait(config)
 		})
 		It("should not set a CPU limit if the namespace doesn't match the selector", func() {
 			By("Creating a running VMI")
@@ -3040,7 +3040,7 @@ var _ = Describe("[sig-compute]Configurations", decorators.SigCompute, func() {
 			// Clear SMBios values if already set in kubevirt-config, for testing default values.
 			test_smbios := &v1.SMBiosConfiguration{Family: "", Product: "", Manufacturer: ""}
 			config.SMBIOSConfig = test_smbios
-			tests.UpdateKubeVirtConfigValueAndWait(config)
+			kvconfig.UpdateKubeVirtConfigValueAndWait(config)
 
 			By("Starting a VirtualMachineInstance")
 			vmi, err := virtClient.VirtualMachineInstance(testsuite.GetTestNamespace(vmi)).Create(context.Background(), vmi, metav1.CreateOptions{})
@@ -3075,7 +3075,7 @@ var _ = Describe("[sig-compute]Configurations", decorators.SigCompute, func() {
 			// Set a custom test SMBios
 			test_smbios := &v1.SMBiosConfiguration{Family: "test", Product: "test", Manufacturer: "None", Sku: "1.0", Version: "1.0"}
 			config.SMBIOSConfig = test_smbios
-			tests.UpdateKubeVirtConfigValueAndWait(config)
+			kvconfig.UpdateKubeVirtConfigValueAndWait(config)
 
 			By("Starting a VirtualMachineInstance")
 			vmi, err := virtClient.VirtualMachineInstance(testsuite.GetTestNamespace(vmi)).Create(context.Background(), vmi, metav1.CreateOptions{})
@@ -3198,7 +3198,7 @@ var _ = Describe("[sig-compute]Configurations", decorators.SigCompute, func() {
 			if testingPciFunctions {
 				assignDisksToFunctions(startIndex, vmi)
 			} else {
-				config.DisableFeatureGate(virtconfig.ExpandDisksGate)
+				kvconfig.DisableFeatureGate(virtconfig.ExpandDisksGate)
 				assignDisksToSlots(startIndex, vmi)
 			}
 			vmi, err := virtClient.VirtualMachineInstance(testsuite.GetTestNamespace(vmi)).Create(context.Background(), vmi, metav1.CreateOptions{})
